Criação do componente ContainerResultCEP
