+++ busca-cep/src/components/ContainerResultCEP/style.css
@@ -23,10 +23,6 @@
   width: 700px;
 }
 
-/* .title {
-  margin-bottom: 1rem;
-} */
-
 .search_container {
   display: flex;
   justify-content: center;
